+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.20.3](https://github.com/adobe/aem-sidekick/compare/v1.20.2...v1.20.3) (2024-04-27)
+
+
+### Bug Fixes
+
+* toast when activating config, improve tests ([#126](https://github.com/adobe/aem-sidekick/issues/126)) ([4551c33](https://github.com/adobe/aem-sidekick/commit/4551c33683e810fb4db51dfefeb831596aaf5cae))
+
+## [1.20.2](https://github.com/adobe/aem-sidekick/compare/v1.20.1...v1.20.2) (2024-04-27)
+
+
+### Bug Fixes
+
+* don't load content script when hidden ([#125](https://github.com/adobe/aem-sidekick/issues/125)) ([68bbf3d](https://github.com/adobe/aem-sidekick/commit/68bbf3dbf454d68dae797d8d9c4a5e4bbbd297e3))
+
 ## [1.20.1](https://github.com/adobe/aem-sidekick/compare/v1.20.0...v1.20.1) (2024-04-26)